requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=3.0.0
bcrypt>=4.0.0
PyJWT>=2.6.0
pydantic>=1.9.0
sqlalchemy>=1.4.0
requests>=2.28.0openai-whisper>=20230314
python-multipart>=0.0.6